rename notrace registration and standardize trace_utils imports
